an initial working version of manticore create indexes, create, and read data

+++ backend/db/manticore_db.py
@@ -1,4 +1,5 @@
 
+from copy import copy
 import json
 import re
 from typing import Any
@@ -7,7 +8,7 @@ import manticoresearch
 from fastapi.logger import logger
 from models.core import EntityDTO, Meta, Space
 from models.enums import LockAction, ResourceType, SortType
-from utils.helpers import resolve_schema_references
+from utils.helpers import delete_none, resolve_schema_references
 from utils.settings import settings
 
 
@@ -30,55 +31,55 @@ class ManticoreDB(BaseDB):
         "subpath": "string",
         "exact_subpath": "string",
         "resource_type": "string",
-        "displayname_en": "string",
-        "displayname_ar": "string",
-        "displayname_kd": "string",
-        "description_en": "string",
-        "description_ar": "string",
-        "description_kd": "string",
+        "displayname": "json",
+        # "displayname_ar": "string",
+        # "displayname_kd": "string",
+        "description": "json",
+        # "description_ar": "string",
+        # "description_kd": "string",
 
         "is_active": "bool",
 
         "payload_content_type": "string",
         "schema_shortname": "string",
-        
-        "payload_content_type": "string",
-        "payload_content_type": "string",
-        
+                
         "created_at": "timestamp",
         "updated_at": "timestamp",
         
         "view_acl": "text",
-        "tags": "text",
-        "query_policies": "text",
+        "tags": "json",
+        "query_policies": "json",
         
         "owner_shortname": "string",
-        "payload_content_type": "string",
-
-
+        
         # User fields
         "msisdn": "string",
         "email": "string",
+        
         # Ticket fields
         "state": "string",
         "is_open": "bool",
         "workflow_shortname": "string",
-        "collaborators_delivered_by": "string",
-        "collaborators_processed_by": "string",
+        "collaborators": "json",
+        # "collaborators_delivered_by": "string",
+        # "collaborators_processed_by": "string",
         "resolution_reason": "string",
 
         # Notification fields
         "type": "string",
         "is_read": "string",
         "priority": "string",
-        "reporter_type": "string",
-        "reporter_name": "string",
-        "reporter_channel": "string",
-        "reporter_distributor": "string",
-        "reporter_governorate": "string",
-        "reporter_msisdn": "string",
+        "reporter": "json",
+        # "reporter_type": "string",
+        # "reporter_name": "string",
+        # "reporter_channel": "string",
+        # "reporter_distributor": "string",
+        # "reporter_governorate": "string",
+        # "reporter_msisdn": "string",
         
         "payload_string": "string",
+        
+        "document_data": "json"
     }
     
     SCHEMA_DATA_TYPES_MAPPER = {
@@ -131,7 +132,8 @@ class ManticoreDB(BaseDB):
 
     async def create_index(self, name: str, fields: dict[str, str], **kwargs) -> bool:
         name = name.replace(":", "__")
-        await self.drop_index(name, False)
+        if self.is_index_exist(name):
+            await self.drop_index(name, False)
         
         try:
             sql_str = f"CREATE TABLE {name}("
@@ -154,8 +156,8 @@ class ManticoreDB(BaseDB):
         """
         takes a property of the json schema definition, and returns the db schema index
         """
-        
-        if not isinstance(property, dict) or key_chain.endswith("."):
+        # pp(key_chain, property, db_schema_definition)
+        if not isinstance(property, dict) or key_chain.endswith("_"):
             return db_schema_definition
 
         if "type" in property and property["type"] != "object":
@@ -178,7 +180,7 @@ class ManticoreDB(BaseDB):
                 ].items():
                     if property_value["type"] != "string":
                         continue
-                    db_schema_definition["{key_chain}_{property_key}"] = "text"
+                    db_schema_definition[f"{key_chain}_{property_key}"] = "text"
                 return db_schema_definition
 
             # if property["type"] == "array":
@@ -199,11 +201,11 @@ class ManticoreDB(BaseDB):
 
         for property_key, property_value in property["properties"].items():
             db_schema_definition = self.get_index_fields_from_json_schema_property(
-                f"{key_chain}.{property_key}", property_value, db_schema_definition
+                f"{key_chain}_{property_key}", property_value, db_schema_definition
             )
 
         return db_schema_definition
-    
+
     async def create_application_indexes(
         self, 
         for_space: str | None = None, 
@@ -213,6 +215,9 @@ class ManticoreDB(BaseDB):
     ) -> None:
         
         spaces = await self.find_by_id("spaces")
+        branch_name = 'master'
+        
+        await self.create_custom_indices()
         for space_name in spaces:
             space_obj = Space.model_validate_json(spaces[space_name])
             if (
@@ -221,7 +226,7 @@ class ManticoreDB(BaseDB):
                 continue
 
             await self.create_index(
-                f"{space_name}:meta", self.META_SCHEMA
+                f"{space_name}:{branch_name}:meta", self.META_SCHEMA
             )
 
             # CREATE DB INDEX FOR EACH SCHEMA DEFINITION INSIDE THE SPACE
@@ -250,7 +255,7 @@ class ManticoreDB(BaseDB):
                 # GENERATE DB INDEX DEFINITION BY MAPPIN SCHEMA PROPERTIES TO DB INDEX FIELDS
                 schema_content = json.loads(schema_path.read_text())
                 schema_content = resolve_schema_references(schema_content)
-                db_schema_definition: dict[str, str] = self.META_SCHEMA
+                db_schema_definition: dict[str, str] = copy(self.META_SCHEMA)
                 if "properties" in schema_content:
                     for key, property in schema_content["properties"].items():
                         generated_schema_fields = self.get_index_fields_from_json_schema_property(
@@ -270,7 +275,7 @@ class ManticoreDB(BaseDB):
                     db_schema_definition["meta_doc_id"] = "string"
 
                     await self.create_index(
-                        f"{space_name}:{schema_shortname}",
+                        f"{space_name}:{branch_name}:{schema_shortname}",
                         db_schema_definition
                     )
 
@@ -308,6 +313,7 @@ class ManticoreDB(BaseDB):
     
     async def create_custom_indices(self, for_space: str | None = None) -> None:
         # redis_schemas: dict[str, list] = {}
+        # branch_name = 'master'
         for i, index in enumerate(self.SYS_INDEXES):
             if (
                 for_space
@@ -327,9 +333,9 @@ class ManticoreDB(BaseDB):
                 index["class"], index["exclude_from_index"]
             )
             
-            generated_schema_fields.update(self.META_SCHEMA)
+            self.META_SCHEMA.update(generated_schema_fields)
             
-            await self.create_index(f"{index['space']}:meta", generated_schema_fields)
+            # await self.create_index(f"{index['space']}:{branch_name}:meta", generated_schema_fields)
 
             # redis_schemas[f"{index['space']}:{index['branch']}"] = (
             #     self.append_unique_index_fields(
@@ -456,15 +462,29 @@ class ManticoreDB(BaseDB):
         
         doc_id_parts = doc_id.split(":")
         
-        if len(doc_id_parts) != 3:
+        if len(doc_id_parts) < 3:
             raise Exception(f"Invalid document id, {doc_id}")
         
-        return f"{doc_id_parts[0]}__{doc_id_parts[1]}"
+        return f"{doc_id_parts[0]}__{doc_id_parts[1]}__{doc_id_parts[2]}"
+    
     
     async def find_by_id(self, id: str) -> dict[str, Any]:
         try:
             index_name: str = self.get_index_name_from_doc_id(id)
-            return {}
+            identifier = "document_id" 
+            if index_name == "key_value_pairs":
+                identifier = "key"
+            find_res = self.mc_command(f"select * from {index_name} where {identifier} = '{id}' limit 1")
+
+            if not find_res or find_res['total'] == 0 or len(find_res["data"]) == 0:
+                return {}
+            
+            if isinstance(find_res["data"][0]["value"], dict):
+                return find_res["data"][0]["value"]
+            elif isinstance(find_res["data"][0]["value"], str):
+                return json.loads(find_res["data"][0]["value"])
+            else:
+                raise Exception(f"Invalid data type {type(find_res['data'][0])}")
         except Exception as e:
             logger.error(f"Error at ManticoreDB.find_by_id: {e.args}")
             return {}
@@ -486,22 +506,131 @@ class ManticoreDB(BaseDB):
         if ":" not in id:
             await self.set_key(id, json.dumps(doc))
             
-        return True
-
+        try:
+            doc['document_id'] = id
+            index = self.get_index_name_from_doc_id(id)
+            self.indexApi.insert({
+                "index": index,
+                "doc": doc
+            })
+            return True
+        except Exception as e:
+            logger.error(f"Error at ManticoreDB.save_at_id: {e}")
+            return False
 
     async def prepare_meta_doc(
         self, space_name: str, branch_name: str | None, subpath: str, meta: Meta
     ) -> tuple[str, dict[str, Any]]:
-        return ("", {})
+        resource_type = meta.__class__.__name__.lower()
+        meta_doc_id = self.generate_doc_id(
+            space_name, branch_name, "meta", meta.shortname, subpath
+        )
+        payload_doc_id = None
+        if meta.payload and meta.payload.schema_shortname:
+            payload_doc_id = self.generate_doc_id(
+                space_name,
+                branch_name,
+                meta.payload.schema_shortname,
+                meta.shortname,
+                subpath,
+            )
+        meta_json: dict[str, Any] = json.loads(meta.model_dump_json(exclude_none=True))
+        meta_json["query_policies"] = self.generate_query_policies(
+            space_name,
+            subpath,
+            resource_type,
+            meta.is_active,
+            meta.owner_shortname,
+            meta.owner_group_shortname,
+            meta.shortname,
+        )
+        meta_json["view_acl"] = self.generate_view_acl(meta_json.get("acl"))
+        meta_json["subpath"] = subpath
+        meta_json["branch_name"] = branch_name
+        meta_json["resource_type"] = resource_type
+        meta_json["created_at"] = meta.created_at.timestamp()
+        meta_json["updated_at"] = meta.updated_at.timestamp()
+        meta_json["payload_doc_id"] = payload_doc_id
+        
+        meta_json = self.filter_unindexed_attributes(meta_json, meta_doc_id)
+        
+        # Encode list to json string
+        meta_json = {key: json.dumps(value) if isinstance(value, list) else value for key, value in meta_json.items()}
 
+        return meta_doc_id, meta_json
 
+    def filter_unindexed_attributes(self, doc: dict[str, Any], id: str) -> dict[str, Any]:
+        # pp(before_filter__doc=doc, id=id)
+        filtered_doc = {
+            'document_data': doc,
+        }
+        
+        index_name = self.get_index_name_from_doc_id(id)
+        res = self.mc_command(f"describe {index_name}")
+        if not res:
+            return filtered_doc
+        
+        index_schema = res['data']
+        index_fields = {field['Field'] for field in index_schema}
+        # pp(index_schema=index_schema)
+        filtered_doc = {field: doc[field] for field in doc if field in index_fields}
+        # pp(after_filter__doc=doc, id=id)
+        return delete_none(filtered_doc)
+    
+    
     async def prepare_payload_doc(
         self,
         dto: EntityDTO,
         meta: Meta,
         payload: dict[str, Any],
     ) -> tuple[str, dict[str, Any]]:
-        return ("", {})
+        if meta.payload is None:
+            raise Exception(
+                f"Missing payload for {dto.space_name}/{dto.branch_name}/{dto.subpath} of type {dto.resource_type}"
+            )
+        if meta.payload.body is None:
+            raise Exception(
+                f"Missing body for {dto.space_name}/{dto.branch_name}/{dto.subpath} of type {dto.resource_type}"
+            )
+        if not isinstance(meta.payload.body, str):
+            raise Exception("body should be type of string")
+        payload_shortname = meta.payload.body.split(".")[0]
+        meta_doc_id = self.generate_doc_id(
+            dto.space_name, dto.branch_name, "meta", payload_shortname, dto.subpath
+        )
+        docid: str = self.generate_doc_id(
+            dto.space_name,
+            dto.branch_name,
+            meta.payload.schema_shortname or "",
+            payload_shortname,
+            dto.subpath,
+        )
+
+        payload["query_policies"] = self.generate_query_policies(
+            dto.space_name,
+            dto.subpath,
+            dto.resource_type or ResourceType.content,
+            meta.is_active,
+            meta.owner_shortname,
+            meta.owner_group_shortname,
+            meta.shortname,
+        )
+        if not payload["query_policies"]:
+            print(
+                f"Warning: this entry `{dto.space_name}/{dto.subpath}/{meta.shortname}` can't be accessed"
+            )
+        payload["subpath"] = dto.subpath
+        payload["branch_name"] = dto.branch_name
+        payload["resource_type"] = dto.resource_type
+        payload["shortname"] = payload_shortname
+        payload["meta_doc_id"] = meta_doc_id
+
+        payload = self.filter_unindexed_attributes(payload, docid)
+        
+        # Encode list to json string
+        payload = {key: json.dumps(value) if isinstance(value, list) else value for key, value in payload.items()}
+        
+        return docid, payload
